+++ CHANGELOG.md
@@ -4,8 +4,16 @@
 
 ### Added
 
+### Modified
+
+- Fix'd description field for pending status response
+
+## [2.0.0]
+### Added
+
 - changelog
 - status and coverage endpoints for shields
+- coverage increase check on pull request events
 
 ### Modified